+++ app/services/hyrax/workflow/sixty_day_embargo_notification.rb
@@ -3,32 +3,15 @@ module Hyrax
     # Notification that embargo will expire sixty days from today.
     # Should notify work depositor only
     class SixtyDayEmbargoNotification
-      include ActionView::Helpers::UrlHelper
-      def self.send_notification(work_id)
-        notification = SixtyDayEmbargoNotification.new(work_id)
-        notification.call
-        notification
-      end
-
-      def initialize(work_id)
-        @work = Etd.find(work_id)
-        @user = ::User.find_by_user_key(@work.depositor)
-        # Set the user email to the post-graduation email in this ETD,
-        # just long enough to send this message. Don't save it, because
-        # that should only happen at graduation. This is a safeguard just
-        # in case the user email on file doesn't match what's in this ETD.
-        @user.email = @work.post_graduation_email.first
-        @message = message
-        @subject = subject
-      end
+      include EmbargoNotificationBehavior
 
       def subject
-        "Embargo on #{@work.title.first} will expire in sixty days"
+        "Embargo on #{title} will expire in sixty days"
       end
 
       def message
         "Dear #{@user.display_name},\n\n" \
-        "The embargo on your dissertation or thesis, #{@work.title.first}, will expire in sixty days.\n" \
+        "The embargo on your dissertation or thesis, #{title}, will expire in sixty days.\n" \
         "Your dissertation or thesis can be accessed in the Emory University ETD Repository at: #{link_to document_url, document_url} \n\n" \
         "If you need your record to remain restricted/embargoed, please contact etd-help@listserv.cc.emory.edu<mailto:etd-help@listserv.cc.emory.edu> with your request. \n\n" \
         "Please do not reply directy to this email.\n\n" \
@@ -38,26 +21,6 @@ module Hyrax
         "will need to contact ProQuest's Author Relations department to request " \
         "they extend your embargo. You can reach them at 1-800-521-0600 ext. 77020 " \
         "or via email at disspub@proquest.com<mailto:disspub@proquest.com>."
-      end
-
-      def call
-        user = ::User.find_or_create_by(uid: WorkflowSetup::NOTIFICATION_OWNER)
-        Rails.logger.warn "ETD #{@work.id}: 60 day embargo expiration notification sent to #{@user.email}"
-        recipients.each do |recipient|
-          user.send_message(recipient, @message, @subject)
-        end
-      end
-
-      # Send this to the depositor only
-      def recipients
-        Array.wrap(@user)
-      end
-
-      # Get the full URL for email notifications
-      # This should get pushed upstream to Hyrax
-      def document_url
-        key = @work.model_name.singular_route_key
-        Rails.application.routes.url_helpers.send(key + "_url", @work.id)
       end
     end
   end
